用户退出群聊

+++ src/main/java/com/cxkj/wechat/netty/executor/single/SingleChatCancelExecutor.java
@@ -1,10 +1,10 @@
 package com.cxkj.wechat.netty.executor.single;
 
-import cn.hutool.core.util.ObjectUtil;
 import com.alibaba.fastjson.JSONObject;
-import com.cxkj.wechat.bo.RequestParamBo;
 import com.cxkj.wechat.constant.Command;
+import com.cxkj.wechat.constant.ResultCodeEnum;
 import com.cxkj.wechat.constant.SystemConstant;
+import com.cxkj.wechat.entity.SingleChat;
 import com.cxkj.wechat.netty.ex.ValidateException;
 import com.cxkj.wechat.netty.executor.ExecutorAnno;
 import com.cxkj.wechat.netty.executor.base.ChatExecutor;
@@ -46,9 +46,14 @@ public class SingleChatCancelExecutor extends ChatExecutor {
     @Override
     protected void concreteAction(Channel channel) {
         Channel userChannel = SessionUtil.getUserChannel(requestParam.getToUserId());
-        if (userChannel != null) {
-            sendMessage(userChannel, JsonResult.success(new SingleChatCancelVo(requestParam.getId()), command));
+        SingleChat singleChat = singleChatService.getById(requestParam.getId());
+        if (singleChat != null && System.currentTimeMillis() < singleChat.getCreateTimes() + SystemConstant.CHAT_CANCEL_TIMES) {
+            if (userChannel != null) {
+                sendMessage(userChannel, JsonResult.success(new SingleChatCancelVo(requestParam.getId()), command));
+            }
+            friendApplicationService.deleteById(requestParam.getId());
+        }else {
+            sendMessage(channel,JsonResult.failed(ResultCodeEnum.UNDO_TIME_EXCEEDED));
         }
-        friendApplicationService.deleteById(requestParam.getId());
     }
 }
